requirements.txt
fastapi>=0.115.6,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
pydantic>=2.10.4,<3.0.0
python-multipart>=0.0.20,<1.0.0
snowflake-connector-python>=3.12.4,<4.0.0
httpx>=0.28.1,<1.0.0
pytest>=8.3.4,<9.0.0
mediapipe
opencv-python-headless
numpy
scikit-learn
google-genai
python-dotenv
Pillow
